\name{cgiarOcs-package}
\alias{cgiarOcs-package}
\alias{cgiarOcs}
\docType{package}
\title{
\packageTitle{cgiarOcs}
}
\description{
\packageDescription{cgiarOcs}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{cgiarOcs}
\packageIndices{cgiarOcs}
~~ An overview of how to use the package, including the most important functions oh yes haha~~
}
\author{
\packageAuthor{cgiarOcs}

Maintainer: \packageMaintainer{cgiarOcs}
}
\references{
~~ Literature or other references for background information ~~
}
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
