\name{cgiarOcs-package}
\alias{cgiarOcs-package}
\alias{cgiarOcs}
\docType{package}
\title{
CGIAR C++ functions
}
\description{

This package aims to store all the functions that heavily rely on the use of the Rcpp package for heavy computations.

}
\details{

This package stores functions for optimal cross selection based on contribution theory.

}
\author{

Author(s): Chris Gaynor (ocs function)

Maintainer: Giovanny Covarrubias-Pazaran

}
\references{

Kinghorn, B. (1999). 19. Mate Selection for the tactical implementation of breeding programs. Proceedings of the Advancement of Animal Breeding and Genetics, 13, 130-133.

Woolliams, J. A., Berg, P., Dagnachew, B. S., & Meuwissen, T. H. E. (2015). Genetic contributions and their optimization. Journal of Animal Breeding and Genetics, 132(2), 89-99.

https://alphagenes.roslin.ed.ac.uk/wp/wp-content/uploads/2019/05/01_OptimalContributionSelection.pdf?x44213

}
\keyword{ package }
\seealso{

Linked functions associated to the package.

}
\examples{
~~ simple examples of the most important functions ~~
}
